+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 8.1
 ---
 
- * Add option `redis_cluster` to force cluster mode
+ * Add option `cluster` to force cluster mode
 
 7.4
 ---